Remove unnecessary check in parser loop
The loop over the class methods we've found included a check to see if
the method extended the appropriate class. Not only was this using
outdated visitor logic but the actual check was being handled before the
loop in such a way that the class methods collection would only include
methods that are extending the TestCase.