Feature/explicit test suite/#14 (#50)
* Parse explicit default test suite

Includes changes necessary to the TestSuiteModel so that we can
determine what class actually backs the TestSuite. Also introduces a
concrete DefaultTestSuite so that we can reference it appropriately as
opposed to an anonymous class that was being used before.

* Barebones TestSuite handled in runner

* Remove unnecessary check in parser loop

The loop over the class methods we've found included a check to see if
the method extended the appropriate class. Not only was this using
outdated visitor logic but the actual check was being handled before the
loop in such a way that the class methods collection would only include
methods that are extending the TestCase.

* Provide ability to specify explicit default test suite