Check that a parsed class is autoloadable
Ensures that a parsed class is autoloadable. Previously if a class that
has a method annotated with an AsyncUnit Attribute could not be loaded
the check for a subclass would fail and the error message would tell you
that you're annotating a method on a class that does not extend the
appropriate type when you were, in fact, extending the appropriate type.
The real problem is that the class extend the AsyncUnit TestCase could
not be autoloaded. This patch ensures that an appropriate error message
is shown stating that you need to setup your autoloader appropriately.