Refactor Parser to return Promise
In preparation for the Parser supporting async I/O we needed to return a
Promise instead of a ParserResult. This change updates the Parser::parse
method signature and updates the appropriate _framework_ code to ensure
that the tests are passing. The cli tests are still expected to fail at
this time. Additional work, including Application dependencies, will
need to be changed to incorporate the changes for async I/O more
completely.